renaming restore component route, impl removing to_delete from component, wip on edge restore
