+++ lib/sdf-server/src/server/service/diagram.rs
@@ -3,6 +3,7 @@ use axum::response::{IntoResponse, Response};
 use axum::routing::{get, post};
 use axum::Json;
 use axum::Router;
+use dal::attribute::prototype::AttributePrototypeError;
 use dal::component::ComponentError;
 use dal::socket::input::InputSocketError;
 use dal::workspace_snapshot::WorkspaceSnapshotError;
@@ -24,9 +25,9 @@ pub mod set_component_position;
 
 // mod connect_component_to_frame;
 pub mod delete_component;
-// pub mod delete_connection;
+//pub mod delete_connection;
 // pub mod paste_component;
-// mod restore_component;
+pub mod remove_delete_intent;
 // pub mod restore_connection;
 
 #[remain::sorted]
@@ -36,6 +37,8 @@ pub enum DiagramError {
     Action(#[from] ActionError),
     #[error("action: {0}")]
     ActionPrototype(#[from] ActionPrototypeError),
+    #[error("prototype error: {0}")]
+    AttributePrototype(#[from] AttributePrototypeError),
     #[error("changeset error: {0}")]
     ChangeSet(#[from] ChangeSetPointerError),
     #[error("change set not found")]
@@ -121,14 +124,10 @@ pub fn routes() -> Router<AppState> {
             "/delete_components",
             post(delete_component::delete_components),
         )
-        // .route(
-        //     "/restore_component",
-        //     post(restore_component::restore_component),
-        // )
-        // .route(
-        //     "/restore_components",
-        //     post(restore_component::restore_components),
-        // )
+        .route(
+            "/remove_delete_intent",
+            post(remove_delete_intent::remove_delete_intent),
+        )
         .route(
             "/connect_component_to_frame",
             post(connect_component_to_frame::connect_component_to_frame),
